Add overrides option
Allows overriding of inclusion vs non inclusion of proposals regardless
of pare option.  Specify review numbers in file with a + or - in front,
and those files will either be always included, or never included,
respectively.

This is tangentially related to #46: A bunch of page formatting and
filtering tweaks.  It was based on a comment there about a weird
proposal that had errors further upstream.